refactor: move validation notifs to own component
