+++ packages/frontend/src/components/NavBar/NotificationsMenu/index.tsx
@@ -1,0 +1,60 @@
+import { Button, Indicator, Menu } from '@mantine/core';
+import { IconBell } from '@tabler/icons-react';
+import { FC } from 'react';
+import {
+    useValidationNotificationChecker,
+    useValidationUserAbility,
+} from '../../../hooks/validation/useValidation';
+import MantineIcon from '../../common/MantineIcon';
+import { ValidationErrorNotification } from './ValidationErrorNotification';
+
+export const NotificationsMenu: FC<{ projectUuid: string }> = ({
+    projectUuid,
+}) => {
+    const canUserManageValidations = useValidationUserAbility(projectUuid);
+    const [hasReadValidationNotification, setHasReadValidationNotification] =
+        useValidationNotificationChecker();
+
+    return (
+        <Menu
+            withArrow
+            shadow="lg"
+            position="bottom-end"
+            arrowOffset={16}
+            offset={-2}
+        >
+            <Menu.Target>
+                <Button
+                    variant="default"
+                    size="xs"
+                    onClick={() => setHasReadValidationNotification()}
+                    sx={{
+                        // NOTE: Revert overflow so badge doesn't get cropped off
+                        '.mantine-Button-label': {
+                            overflow: 'revert',
+                        },
+                    }}
+                >
+                    <Indicator
+                        color="red"
+                        offset={2}
+                        disabled={
+                            !canUserManageValidations ||
+                            hasReadValidationNotification
+                        }
+                    >
+                        <MantineIcon icon={IconBell} />
+                    </Indicator>
+                </Button>
+            </Menu.Target>
+
+            <Menu.Dropdown>
+                {canUserManageValidations ? (
+                    <ValidationErrorNotification projectUuid={projectUuid} />
+                ) : (
+                    <Menu.Item>No notifications</Menu.Item>
+                )}
+            </Menu.Dropdown>
+        </Menu>
+    );
+};
